% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pnad.R
\name{load_pnad}
\alias{load_pnad}
\title{Function to download PNAD (people) data}
\usage{
load_pnad(year)
}
\arguments{
\item{year:}{Year for which you want to download PNAD}
}
\description{
Function to download PNAD (people) data
}
\examples{
load_pnad(2002) # returns PNAD 2002
load_pnad(2015) # returns PNAD 2015

}
\author{
Marcos Júnio Ribeiro
}
