% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_pnad.R
\name{load_pnad}
\alias{load_pnad}
\title{Pnad download}
\usage{
load_pnad(year, type)
}
\arguments{
\item{year:}{Numeric (Positive) year for which you want to download PNAD.}

\item{type:}{A character string (pessoas or domicilios) that indicates which PNAD do you want.}
}
\description{
Returns a tibble with PNAD data for the requested year.
}
\examples{
load_pnad(2002, type='pessoas') # returns PNAD pessoas 2002.
load_pnad(2015, type = 'domicilios') # returns PNAD domicilios 2015.
}
\author{
Marcos Júnio Ribeiro
}
